requirements.txt
cffi==1.7.0
Flask==0.11.1
Flask-PyMongo==0.4.1
Flask-Uploads==0.2.1
jpegtran-cffi==0.5.2
pymongo==3.3.0
requests==2.11.1
